Rework CAGE; add CAGE to root settings view
