+++ Localization/Localizable.xcstrings
@@ -3431,6 +3431,18 @@
         }
       }
     },
+    "Cannula Age" : {
+      "comment" : "Text for cannula age (CAGE)",
+      "extractionState" : "manual",
+      "localizations" : {
+        "en" : {
+          "stringUnit" : {
+            "state" : "translated",
+            "value" : "Cannula Age"
+          }
+        }
+      }
+    },
     "Change your Patch now. Insulin delivery will stop in {0} or when no more insulin remains." : {
       "comment" : "description for grace period notice",
       "extractionState" : "manual",
@@ -6727,18 +6739,6 @@
           "stringUnit" : {
             "state" : "translated",
             "value" : "完成"
-          }
-        }
-      }
-    },
-    "Elapsed time" : {
-      "comment" : "Text for elapsed patch lifetime",
-      "extractionState" : "manual",
-      "localizations" : {
-        "en" : {
-          "stringUnit" : {
-            "state" : "translated",
-            "value" : "Elapsed time"
           }
         }
       }
