+++ MedtrumKitUI/Views/Settings/MedtrumKitSettings.swift
@@ -295,6 +295,19 @@ struct MedtrumKitSettings: View {
             }
 
             Section {
+                HStack {
+                        Text(LocalizedString("Cannula Age", comment: "Text for cannula age (CAGE)"),)
+                            .foregroundColor(Color.primary)
+                        Spacer()
+                        if viewModel.patchLifecycleState != .noPatch {
+                            Text(viewModel.patchLifetime)
+                                .foregroundColor(.secondary)
+                                .multilineTextAlignment(.trailing)
+                        } else {
+                            Text("-")
+                                .foregroundColor(.secondary)
+                        }
+                    }
                 if let activatedAt = viewModel.patchActivatedAt {
                     HStack {
                         Text(LocalizedString("Activation", comment: "Text for activatedAt"))
