+++ MedtrumKitUI/Views/Settings/PatchDetailsView.swift
@@ -35,7 +35,7 @@ struct PatchDetailsView: View {
                     value: viewModel.activatedAt
                 )
                 sectionItem(
-                    title: LocalizedString("Elapsed time", comment: "Text for elapsed patch lifetime"),
+                    title: LocalizedString("Cannula Age", comment: "Text for cannula age (CAGE)"),
                     value: viewModel.patchLifetime
                 )
